Candidate Utilities

NaiveGAflux.CandidateModel — Type

CandidateModel <: Candidate
CandidateModel(model)

A candidate model which can be accessed by model(c) for CandidateModel c.

NaiveGAflux.CandidateOptModel — Type

CandidateOptModel <: AbstractCandidate
CandidateOptModel(candidate::AbstractCandidate, optimizer)

A candidate adding an optimizer to another candidate. The optimizer is accessed by [opt(c)] for CandidateOptModel c.

NaiveGAflux.FittedCandidate — Type

FittedCandidate{F, C} <: AbstractWrappingCandidate
FittedCandidate(c::AbstractCandidate, fitnessvalue, generation)
FittedCandidate(c::AbstractCandidate, fitnessfun::AbstractFitness, generation)

An AbstractCandidate with a computed fitness value. Will compute the fitness value if provided an AbstractFitness.

Basically a container for results so that fitness does not need to be recomputed to e.g. check stopping conditions.

Also useful for fitness smoothing, e.g. with EwmaFitness as it gives access to previous fitness value.

NaiveGAflux.evolvemodel — Function

evolvemodel(m::AbstractMutation{CompGraph}, mapothers=deepcopy)
evolvemodel(m::AbstractMutation{CompGraph}, om::AbstractMutation{FluxOptimizer}, mapothers=deepcopy)

Return a function which maps a AbstractCandidate c1 to a new AbstractCandidate c2 where any CompGraphs g in c1 will be m(deepcopy(g))inc2. Same principle is applied to any optimisers ifom` is present.

All other fields are mapped through the function mapothers (default deepcopy).

Intended use is together with EvolveCandidates.

evolvemodel(m::AbstractCrossover{CompGraph}, mapothers1=deepcopy, mapothers2=deepcopy)
evolvemodel(m::AbstractCrossover{CompGraph}, om::AbstractCrossover{FluxOptimizer}, mapothers1=deepcopy, mapothers2=deepcopy)

Return a function which maps a tuple of AbstractCandidates (c1,c2) to two new candidates c1', c2' where any CompGraphs g1 and g2 in c1 and c2 respectively will be g1', g2' = m((deepcopy(g1), deepcopy(g2))) in c1' and c2' respectively. Same principle applies to any optimisers if om is present.

All other fields in c1 will be mapped through the function mapothers1 and likewise for c2 and mapothers2.

Intended use is together with PairCandidates and EvolveCandidates.

NaiveGAflux.Population — Type

struct Population{N, P}
Population(gen, members)
Population(members)

Basic population type which just adds generation counting to its members.

Evolving the population returns a new Population with the generation counter incremented.

NaiveGAflux.model — Function

model(c::AbstractCandidate; [default])

Return the model of candidate c if c has a model, default (which defaults to nothing) otherwise.

model(f, c::AbstractCandidate; [default])

Return the result of f([model(c; default)]).

NaiveGAflux.opt — Function

opt(c::AbstractCandidate; [default])

Return the optimizer of candidate c if c has an optimizer, default (which defaults to nothing) otherwise.

NaiveGAflux.lossfun — Function

lossfun(c::AbstractCandidate; [default])

Return the loss function of candidate c if c has a lossfunction, default (which defaults to nothing) otherwise.