Misc. Utilities

NaiveGAflux.MutationShield — Type

MutationShield <: DecoratingTrait
MutationShield(t, allowed...)

Shields its associated vertex from being selected for mutation.

Any types in allowed will be allowed to mutate the vertex if supplied when calling allow_mutation.

Note that vertex might still be modified if an adjacent vertex is mutated in a way which propagates to a shielded vertex.

NaiveGAflux.ApplyIf — Type

ApplyIf <: DecoratingTrait
ApplyIf(predicate::Function, apply::Function, base::MutationTrait)

Enables calling apply(v) for an AbstractVertex v which has this trait if 'predicate(v) == true'.

Motivating use case is to have a way to remove vertices which have ended up as noops, e.g. element wise and concatenation vertices with a single input or identity activation functions.

NaiveGAflux.PersistentArray — Type

PersistentArray{T, N} <: AbstractArray{T, N}
PersistentArray(savedir::String, nr::Integer, generator;suffix=".jls")
PersistentArray(savedir::String, suffix::String, data::Array)

Simple persistent array. Can be created from serialized data and can be asked to persist its elements using persist.

Note that once initialized, the array is not backed by the serialized data. Adding/deleting files is not reflected in data and vice versa.

NaiveGAflux.persist — Function

persist(a::PersistentArray)

Serializes the elements of a, one file per element.

NaiveGAflux.ShieldedOpt — Type

ShieldedOpt{O} <: Flux.Optimise.AbstractOptimiser 
ShieldedOpt(o)

Shields o from mutation by OptimizerMutation.